Move FooChooserFactory definitions into the widget JS endpoints rather than the telepath ones
This makes it cleaner for FooChooser and FooChooserFactory to share definitions (e.g. modal handler classes)
