+++ client/src/entrypoints/admin/page-chooser-telepath.js
@@ -1,7 +1,4 @@
-import { ChooserFactory } from '../../components/ChooserWidget';
-
-class PageChooserFactory extends ChooserFactory {
-  // eslint-disable-next-line no-undef
-  widgetClass = PageChooser;
-}
-window.telepath.register('wagtail.widgets.PageChooser', PageChooserFactory);
+window.telepath.register(
+  'wagtail.widgets.PageChooser',
+  window.PageChooserFactory,
+);
